+++ api/categories/delete.php
@@ -6,12 +6,12 @@ header('Content-Type: application/json');*/
 header('Access-Control-Allow-Methods: DELETE');
 header('Access-Control-Allow-Headers: Access-Control-Allow-Headers, Content-Type, Access-Control-Allow-Methods, Authorization, X-Requested-With');
 header('Content-Type: application/json');
-/*
+
 include_once '../../config/Database.php';
 include_once '../../models/Category.php';
 
 $database = new Database();
-$db = $database->connect();*/
+$db = $database->connect();
 
 //Instantiate category object
 $category = new Category($db);
@@ -19,14 +19,19 @@ $category = new Category($db);
 //Get raw posted data
 $data = json_decode(file_get_contents("php://input"));
 
-$category->id = $data->id;
-
-//Create category
-if($category->delete())
-{
-    echo json_encode(array('message' => 'Category Deleted'));
+if(!get_object_vars($data) || !isset($data->id)){ //If there are no parameters
+    echo json_encode(array('message' => 'Missing Required Parameters'));
 }
-else
-{
-    echo json_encode(array('message' => 'Category Not Deleted'));
+else{
+    $category->id = $data->id;
+
+    //Create category
+    if($category->delete())
+    {
+        echo json_encode(array('message' => 'Category Deleted'));
+    }
+    else
+    {
+        echo json_encode(array('message' => 'Category Not Deleted'));
+    }
 }
